Make it possible to link crosswalk against libchromium-ewk.so
As it is today, on regular chromium builds (i.e. non component-build),
'xwalk' (the binary) links against lots of static internal chromium
libs (.a files). That produces a large binary size (~90Mb).
This is because of the way targets' dependencies are set in crosswalk
GYP files.

Patch adds some GYP variables to act as placeholders, and then make
it possible to "remove" some dependencies set in crosswalk's GYP files.
Additionally, it allows 'xwalk' (the binary) to link against libchromium-efl.so,
instead of again internal chromium libs.

We then remove duplicated symbols (in xwalk statically linked binary and
in libchromium-efl.so), and reduce xwalk binary.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=13316
Signed-off-by: Antonio Gomes <a1.gomes@samsung.com>